+++ src/core/io/lottie/lottie_exporter.hpp
@@ -21,8 +21,13 @@ class LottieExporterState
     static constexpr const char* version = "5.7.1";
 
 public:
-    explicit LottieExporterState(ImportExport* format, model::Document* document, bool strip, bool strip_raster )
-        : format(format), document(document), strip(strip), strip_raster( strip_raster ) {}
+    explicit LottieExporterState(ImportExport* format, model::Document* document, bool strip, bool strip_raster, const QVariantMap& settings )
+        : format(format),
+        document(document),
+        strip(strip),
+        strip_raster( strip_raster ),
+        auto_embed(settings["auto_embed"].toBool())
+    {}
 
     QCborMap to_json()
     {
@@ -575,7 +580,18 @@ public:
         if ( !strip_raster )
         {
             for ( const auto& bmp : document->assets()->images->values )
-                assets.push_back(convert_bitmat(bmp.get()));
+            {
+                if ( auto_embed && !bmp->embedded() )
+                {
+                    auto clone = bmp->clone_covariant();
+                    clone->embed(true);
+                    assets.push_back(convert_bitmat(clone.get()));
+                }
+                else
+                {
+                    assets.push_back(convert_bitmat(bmp.get()));
+                }
+            }
         }
 
         for ( const auto& comp : document->assets()->precompositions->values )
@@ -683,6 +699,7 @@ public:
     app::log::Log logger{"Lottie Export"};
     model::Layer* mask = 0;
     bool strip_raster;
+    bool auto_embed;
 };